chore(refactor): set constants once to make sure single I/O hits
