+++ src/commands/compile/internal/getAllJobFolders.ts
@@ -1,5 +1,4 @@
 import { Target } from '@sasjs/utils'
-import { getConstants } from '../../../constants'
 import { getLocalOrGlobalConfig } from '../../../utils/config'
 import { getAbsolutePath } from '../../../utils/utils'
 
@@ -15,7 +14,7 @@ export async function getAllJobFolders(target: Target) {
 
   allJobs = allJobs.filter((s) => !!s) as string[]
 
-  const { buildSourceFolder } = await getConstants()
+  const { buildSourceFolder } = process.sasjsConstants
   allJobs = allJobs.map((job) => getAbsolutePath(job, buildSourceFolder))
 
   return [...new Set(allJobs)]
