+++ src/commands/deploy/spec/getDeployScripts.spec.ts
@@ -7,7 +7,6 @@ import {
   removeTestApp
 } from '../../../utils/test'
 import { getDeployScripts } from '../internal/getDeployScripts'
-import { getConstants } from '../../../constants'
 
 describe('getDeployScripts', () => {
   let target: Target
@@ -31,7 +30,7 @@ describe('getDeployScripts', () => {
   it('should include the build output file for SAS9 when deployServicePack is true', async () => {
     target.buildConfig!.buildOutputFileName = 'output-test.sas'
     const deployScripts = await getDeployScripts(target)
-    const { buildDestinationFolder } = await getConstants()
+    const { buildDestinationFolder } = process.sasjsConstants
 
     expect(deployScripts).toIncludeAllMembers([
       path.join(buildDestinationFolder, 'output-test.sas')
@@ -41,7 +40,7 @@ describe('getDeployScripts', () => {
   it('should should fallback to target name for the build output file name for SAS9', async () => {
     target.buildConfig!.buildOutputFileName = ''
     const deployScripts = await getDeployScripts(target)
-    const { buildDestinationFolder } = await getConstants()
+    const { buildDestinationFolder } = process.sasjsConstants
 
     expect(deployScripts).toIncludeAllMembers([
       path.join(buildDestinationFolder, `${target.name}.sas`)
