+++ src/commands/flow/spec/flow.spec.server.ts
@@ -30,7 +30,6 @@ import {
   LogLevel,
   generateTimestamp
 } from '@sasjs/utils'
-import { getConstants } from '../../../constants'
 import SASjs from '@sasjs/adapter/node'
 
 describe('sasjs flow', () => {
@@ -449,7 +448,7 @@ const createGlobalTarget = async (serverType = ServerType.SasViya) => {
       ? process.env.VIYA_SERVER_URL
       : process.env.SAS9_SERVER_URL) as string,
     appLoc: `/Public/app/cli-tests/${targetName}`,
-    contextName: (await getConstants()).contextName,
+    contextName: process.sasjsConstants.contextName,
     serviceConfig: {
       serviceFolders: ['sasjs/testServices', 'sasjs/testJob', 'sasjs/services'],
       initProgram: 'sasjs/testServices/serviceinit.sas',
