+++ src/commands/folder/spec/folder.spec.server.ts
@@ -6,7 +6,6 @@ import {
   saveToGlobalConfig
 } from '../../../utils/config'
 import { Command } from '../../../utils/command'
-import { getConstants } from '../../../constants'
 
 describe('sasjs folder operations', () => {
   let target: Target
@@ -208,7 +207,7 @@ const createGlobalTarget = async (serverType = ServerType.SasViya) => {
     name: targetName,
     appLoc: `/Public/app/cli-tests/${targetName}`,
     serverType,
-    contextName: (await getConstants()).contextName,
+    contextName: process.sasjsConstants.contextName,
     serverUrl: (serverType === ServerType.SasViya
       ? process.env.VIYA_SERVER_URL
       : process.env.SAS9_SERVER_URL) as string,
