+++ src/commands/run/run.ts
@@ -13,7 +13,6 @@ import {
 } from '@sasjs/utils'
 import { Command } from '../../utils/command'
 import { displayError } from '../../utils/displayResult'
-import { getConstants } from '../../constants'
 import { compileSingleFile } from '../'
 import { displaySasjsRunnerError, getAbsolutePath } from '../../utils/utils'
 import axios from 'axios'
@@ -38,7 +37,7 @@ export async function runSasCode(command: Command) {
     await axios
       .get(filePath)
       .then(async (res) => {
-        const { invalidSasError } = await getConstants()
+        const { invalidSasError } = process.sasjsConstants
         if (typeof res.data !== 'string') {
           throw new Error(invalidSasError)
         }
@@ -152,7 +151,7 @@ async function executeOnSasViya(
 
   process.logger?.success('Job execution completed!')
 
-  const { buildDestinationResultsFolder } = await getConstants()
+  const { buildDestinationResultsFolder } = process.sasjsConstants
   process.logger?.info(
     `Creating ${
       isOutput ? 'output' : 'log'
@@ -187,7 +186,7 @@ async function executeOnSas9(
   }
 
   if (!username || !password) {
-    const { sas9CredentialsError } = await getConstants()
+    const { sas9CredentialsError } = process.sasjsConstants
     throw new Error(sas9CredentialsError)
   }
 
@@ -200,7 +199,7 @@ async function executeOnSas9(
     serverType: target.serverType,
     debug: true
   })
-  const { buildDestinationResultsFolder } = await getConstants()
+  const { buildDestinationResultsFolder } = process.sasjsConstants
   const executionResult = await sasjs
     .executeScriptSAS9(linesToExecute, username, password)
     .catch(async (err) => {
@@ -239,7 +238,7 @@ async function executeOnSas9(
 
 async function createOutputFile(log: string) {
   const timestamp = generateTimestamp()
-  const { buildDestinationResultsFolder } = await getConstants()
+  const { buildDestinationResultsFolder } = process.sasjsConstants
   const outputFilePath = path.join(
     buildDestinationResultsFolder,
     `sasjs-run-${timestamp}.log`
