+++ src/commands/run/spec/run.spec.server.ts
@@ -20,7 +20,6 @@ import {
   updateConfig
 } from '../../../utils/test'
 import { build } from '../../build/build'
-import { getConstants } from '../../../constants'
 
 describe('sasjs run', () => {
   let target: Target
@@ -70,7 +69,7 @@ describe('sasjs run', () => {
 
     it('should throw an error if url response starts with angular bracket(<)', async () => {
       const url = 'https://github.com/sasjs/cli/issues/808'
-      const { invalidSasError } = await getConstants()
+      const { invalidSasError } = process.sasjsConstants
       const error = new Error('Error: ' + invalidSasError)
       await expect(
         runSasCode(new Command(`run -t ${target.name} ${url}`))
@@ -79,7 +78,7 @@ describe('sasjs run', () => {
 
     it('should throw an error when url response is not a string', async () => {
       const url = 'https://api.agify.io/?name=sabir'
-      const { invalidSasError } = await getConstants()
+      const { invalidSasError } = process.sasjsConstants
       const error = new Error('Error: ' + invalidSasError)
       await expect(
         runSasCode(new Command(`run -t ${target.name} ${url}`))
