+++ src/types/system/process.d.ts
@@ -3,5 +3,6 @@ declare namespace NodeJS {
     projectDir: string
     currentDir: string
     logger: import('@sasjs/utils/logger').Logger
+    sasjsConstants: import('../../constants').Constants
   }
 }
