+++ src/utils/index.ts
@@ -5,5 +5,6 @@ export * from './displayResult'
 export * from './file'
 export * from './loadEnvVariables'
 export * from './setProjectDir'
+export * from './setConstants'
 export * from './utils'
 export * from './validateTargetName'
